test: mv server test.
