+++ internal/server_test/adaptor_test.go
@@ -1,4 +1,4 @@
-package tests
+package server
 
 import (
 	"encoding/json"
