+++ internal/server_test/service_selection_test.go
@@ -1,4 +1,4 @@
-package tests
+package server
 
 import (
 	"testing"
